Refactor API and UI structure; rename files and consolidate model configurations
